+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/autocrafter/InWorldExternalPatternSinkKey.java
@@ -1,8 +1,0 @@
-package com.refinedmods.refinedstorage.common.autocrafting.autocrafter;
-
-import com.refinedmods.refinedstorage.api.autocrafting.task.ExternalPatternSinkKey;
-
-import net.minecraft.world.item.ItemStack;
-
-public record InWorldExternalPatternSinkKey(String name, ItemStack stack) implements ExternalPatternSinkKey {
-}