+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/monitor/AutocraftingMonitorItemTooltip.java
@@ -3,7 +3,7 @@ package com.refinedmods.refinedstorage.common.autocrafting.monitor;
 import com.refinedmods.refinedstorage.api.autocrafting.status.TaskStatus;
 import com.refinedmods.refinedstorage.common.api.RefinedStorageClientApi;
 import com.refinedmods.refinedstorage.common.api.support.resource.ResourceRendering;
-import com.refinedmods.refinedstorage.common.autocrafting.autocrafter.InWorldExternalPatternSinkKey;
+import com.refinedmods.refinedstorage.common.autocrafting.autocrafter.AutocrafterExternalPatternSinkKey;
 
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.Font;
@@ -61,15 +61,10 @@ class AutocraftingMonitorItemTooltip implements ClientTooltipComponent {
             );
             yy += 9 + SPACING;
         }
-        if (item.sinkKey() instanceof InWorldExternalPatternSinkKey(String name, ItemStack stack)) {
+        if (item.sinkKey() instanceof AutocrafterExternalPatternSinkKey(_, String name, ItemStack stack)
+            && name != null && stack != null) {
             graphics.item(stack, x, yy);
-            graphics.text(
-                font,
-                name,
-                x + 18 + SPACING,
-                yy + 4,
-                0xFFAAAAAA
-            );
+            graphics.text(font, name, x + 18 + SPACING, yy + 4, 0xFFAAAAAA);
         }
     }
 
@@ -84,10 +79,16 @@ class AutocraftingMonitorItemTooltip implements ClientTooltipComponent {
     public int getWidth(final Font font) {
         final int resourceWidth = font.width(rendering.getDisplayName(item.resource()));
         final int errorWidth = item.type() != TaskStatus.ItemType.NORMAL ? font.width(getErrorTooltip(item.type())) : 0;
-        final int sinkWidth = item.sinkKey() instanceof InWorldExternalPatternSinkKey sinkKey
-            ? (18 + SPACING + font.width(sinkKey.name()))
-            : 0;
+        final int sinkWidth = item.sinkKey() instanceof AutocrafterExternalPatternSinkKey key ? getWidth(font, key) : 0;
         return Math.max(resourceWidth, Math.max(errorWidth, sinkWidth));
+    }
+
+    private static int getWidth(final Font font, final AutocrafterExternalPatternSinkKey key) {
+        final String name = key.name();
+        if (name == null) {
+            return 0;
+        }
+        return 18 + SPACING + font.width(name);
     }
 
     private Component getErrorTooltip(final TaskStatus.ItemType type) {