+++ refinedstorage-network/src/main/java/com/refinedmods/refinedstorage/api/network/impl/node/patternprovider/PatternProviderListener.java
@@ -1,5 +1,6 @@
 package com.refinedmods.refinedstorage.api.network.impl.node.patternprovider;
 
+@FunctionalInterface
 public interface PatternProviderListener {
     void receivedExternalIteration();
 }